Added EditProfileScreen with viewmodel + factory
Added route, navigation and composable to NavGraph to EditProfileScreen, new viewModel to handle logic and factory for it, changed AppTopBar icons order so left is account details and right is home and added new methods to userRepository to update nickname, email, password and updateProfilePicture. Added camera and gallery functionality to change profile picture, everything using userPreferences. Design still not good, but functional.